+++ Excel_PRIME/Implementation/Sheet.cs
@@ -13,18 +13,18 @@ namespace ExcelPRIME.Implementation;
 internal sealed class Sheet : ISheetAsync
 {
     private bool _isDisposed;
-    private readonly XmlReaderHelpersAsync _xmlReaderHelper;
+    private readonly IOpenXmlReaderHelpersAsync _xmlReaderHelper;
     private readonly InstanceContext _instanceContext;
     private readonly XmlNameTable _sharedNameTable;
     private readonly Stream _stream;
-    private IXmlSheetReaderAsync? _sheetReader;
+    private IOpenXmlSheetReaderAsync? _sheetReader;
 
     /// <summary>
     /// Get the internal file name of this worksheet
     /// </summary>
     internal static string GetFileName(int index) => $"xl/worksheets/sheet{index}.xml";
 
-    internal Sheet(Stream stream, XmlReaderHelpersAsync xmlReaderHelper, string name, int index, InstanceContext instanceContext)
+    internal Sheet(Stream stream, IOpenXmlReaderHelpersAsync xmlReaderHelper, string name, int index, InstanceContext instanceContext)
     {
         _stream = stream;
         _xmlReaderHelper = xmlReaderHelper;
@@ -228,7 +228,7 @@ internal sealed class Sheet : ISheetAsync
                 }
             }
 
-            _sheetReader = _xmlReaderHelper.CreateSheetReader(_stream, _instanceContext, _sharedNameTable, ct);
+            _sheetReader = (IOpenXmlSheetReaderAsync)_xmlReaderHelper.CreateSheetReader(_stream, _instanceContext, _sharedNameTable, ct);
         }
         while (_sheetReader.CurrentRow < startRow
                && !ct.IsCancellationRequested)